mw/kubernetes: move fallthrough tests out
Remove the testcase duplication and put fallthrough tests in separate
file.

Also make some names shorter and more descriptive.
